app/api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=1.0.0
chromadb>=0.4.13
pyaudio>=0.2.13  # Optionnel - nécessite PortAudio
sounddevice>=0.4.6  # Alternative à pyaudio 
soundfile>=0.12.1  # Pour la manipulation audio
SpeechRecognition>=3.10.0
PyMuPDF>=1.22.3  # Pour la lecture de PDF
ebooklib>=0.18  # Pour la lecture d'ebooks
beautifulsoup4>=4.12.2  # Pour le parsing HTML 

# Knowledge Ingestion Dependencies
spacy>=3.7.0  # NLP processing and entity recognition
google-api-python-client>=2.100.0  # Google Drive API
google-auth-httplib2>=0.2.0  # Google authentication
google-auth-oauthlib>=1.0.0  # Google OAuth2 flow
dropbox>=11.36.0  # Dropbox API integration
requests>=2.31.0  # HTTP requests for APIs
aiohttp>=3.8.0  # Async HTTP client
aiofiles>=23.0.0  # Async file operations
pillow>=10.0.0  # Image processing for OCR
pytesseract>=0.3.10  # OCR engine
scikit-learn>=1.3.0  # ML algorithms for content classification
numpy>=1.24.0  # Numerical computations
pandas>=2.0.0  # Data manipulation
sqlalchemy>=2.0.0  # Database ORM
alembic>=1.12.0  # Database migrations
psutil>=5.9.0  # System monitoring
opencv-python>=4.8.0  # Computer vision for image processing
python-docx>=0.8.11  # Microsoft Word document processing
python-pptx>=0.6.21  # PowerPoint processing
openpyxl>=3.1.0  # Excel file processing